// src_frontend/api-client/models/DeliveryDayAdjustment.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface DeliveryDayAdjustment
 */
export interface DeliveryDayAdjustment {
    /**
     * 
     * @type {string}
     * @memberof DeliveryDayAdjustment
     */
    id?: string;
    /**
     * 
     * @type {number}
     * @memberof DeliveryDayAdjustment
     */
    calendarWeek: number;
    /**
     * 
     * @type {number}
     * @memberof DeliveryDayAdjustment
     */
    adjustedWeekday: number;
    /**
     * 
     * @type {string}
     * @memberof DeliveryDayAdjustment
     */
    growingPeriod: string;
}

/**
 * Check if a given object implements the DeliveryDayAdjustment interface.
 */
export function instanceOfDeliveryDayAdjustment(value: object): value is DeliveryDayAdjustment {
    if (!('calendarWeek' in value) || value['calendarWeek'] === undefined) return false;
    if (!('adjustedWeekday' in value) || value['adjustedWeekday'] === undefined) return false;
    if (!('growingPeriod' in value) || value['growingPeriod'] === undefined) return false;
    return true;
}

export function DeliveryDayAdjustmentFromJSON(json: any): DeliveryDayAdjustment {
    return DeliveryDayAdjustmentFromJSONTyped(json, false);
}

export function DeliveryDayAdjustmentFromJSONTyped(json: any, ignoreDiscriminator: boolean): DeliveryDayAdjustment {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'calendarWeek': json['calendar_week'],
        'adjustedWeekday': json['adjusted_weekday'],
        'growingPeriod': json['growing_period'],
    };
}

  export function DeliveryDayAdjustmentToJSON(json: any): DeliveryDayAdjustment {
      return DeliveryDayAdjustmentToJSONTyped(json, false);
  }

  export function DeliveryDayAdjustmentToJSONTyped(value?: DeliveryDayAdjustment | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'calendar_week': value['calendarWeek'],
        'adjusted_weekday': value['adjustedWeekday'],
        'growing_period': value['growingPeriod'],
    };
}